refactoring: Moved PaginationFactory
