+++ src/Controller/HospitalController.php
@@ -7,6 +7,7 @@ use App\Domain\Command\Hospital\EditHospitalCommand;
 use App\Domain\Contracts\HospitalInterface;
 use App\Domain\Contracts\UserInterface;
 use App\Entity\Hospital;
+use App\Factory\PaginationFactory;
 use App\Form\HospitalType;
 use App\Repository\AllocationRepository;
 use App\Repository\DispatchAreaRepository;
@@ -24,7 +25,6 @@ use App\Service\Filters\SizeFilter;
 use App\Service\Filters\StateFilter;
 use App\Service\Filters\SupplyAreaFilter;
 use App\Service\FilterService;
-use App\Service\PaginationFactory;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
