+++ src/Controller/ImportController.php
@@ -9,6 +9,7 @@ use App\Domain\Command\Import\ImportDataCommand;
 use App\Domain\Contracts\UserInterface;
 use App\Domain\Event\Import\ImportFailedEvent;
 use App\Entity\Import;
+use App\Factory\PaginationFactory;
 use App\Form\ImportType;
 use App\Repository\ImportRepository;
 use App\Service\Filters\HospitalFilter;
@@ -20,7 +21,6 @@ use App\Service\Filters\PageFilter;
 use App\Service\Filters\SearchFilter;
 use App\Service\Filters\UserFilter;
 use App\Service\FilterService;
-use App\Service\PaginationFactory;
 use App\Service\UploadService;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
