+++ src/Controller/Settings/HospitalController.php
@@ -7,6 +7,7 @@ use App\Domain\Command\Hospital\DeleteHospitalCommand;
 use App\Domain\Command\Hospital\EditHospitalCommand;
 use App\Domain\Contracts\HospitalInterface;
 use App\Entity\Hospital;
+use App\Factory\PaginationFactory;
 use App\Form\HospitalType;
 use App\Repository\HospitalRepository;
 use App\Service\Filters\DispatchAreaFilter;
@@ -21,7 +22,6 @@ use App\Service\Filters\SizeFilter;
 use App\Service\Filters\StateFilter;
 use App\Service\Filters\SupplyAreaFilter;
 use App\Service\FilterService;
-use App\Service\PaginationFactory;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
