+++ src/Controller/Settings/UserController.php
@@ -6,6 +6,7 @@ use App\Domain\Command\User\CreateUserCommand;
 use App\Domain\Command\User\EditUserCommand;
 use App\Domain\Command\User\PromoteUserCommand;
 use App\Entity\User;
+use App\Factory\PaginationFactory;
 use App\Form\User\UserType;
 use App\Repository\UserRepository;
 use App\Service\Filters\OrderFilter;
@@ -13,7 +14,6 @@ use App\Service\Filters\PageFilter;
 use App\Service\Filters\SearchFilter;
 use App\Service\FilterService;
 use App\Service\MailerService;
-use App\Service\PaginationFactory;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
