+++ src/Factory/PaginationFactory.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Service;
+namespace App\Factory;
 
 use App\DataTransferObjects\PaginationDto;
